refactor: settings, middlewares 폴더 이름 단수로 변경
