+++ packages/backend/src/controller/problemController.ts
@@ -3,7 +3,7 @@ import express from 'express';
 import { writeProblem } from '../service/problemService';
 import { findProblemCodeById } from '../service/problemCodeService';
 
-import { isLogin } from '../middlewares/isLogin';
+import { isLogin } from '../middleware/isLogin';
 
 const router = express.Router();
